requirements.txt
astroid==1.5.3
autopep8==1.3.3
diff-match-patch==20121119
Django==1.11.8
django-auditlog==0.4.4
django-autocomplete-light==3.2.10
django-bootstrap-datepicker==1.2.3
django-countries==5.0
django-filter==1.1.0
django-import-export==0.6.1
django-jsonfield==1.0.1
django-polymorphic==1.3
django-summernote==0.8.8.4
django-tables2==1.16.0
django-tinymce==2.6.0
django-widget-tweaks==1.4.1
et-xmlfile==1.0.1
isort==4.2.15
jdcal==1.3
lazy-object-proxy==1.3.1
mccabe==0.6.1
odfpy==1.3.5
olefile==0.44
openpyxl==2.4.9
Pillow==5.0.0
psycopg2==2.7.3.2
pycodestyle==2.3.1
pylint==1.7.4
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
reportlab==3.4.0
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
wrapt==1.10.11
xlrd==1.1.0
xlwt==1.3.0
